+++ database/migrations/2023_06_06_000008_create_money_estimates_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->double('min');
             $table->double('max');
-            $table->unsignedBigInteger('weight');
+            $table->unsignedSmallInteger('weight');
             $table->timestamps();
         });
     }
